device bug

+++ train.py
@@ -152,7 +152,7 @@ def weighted_train(epoch, train_loader, p_net, p_optimizer, g_net, g_optimizer, 
             oh = data.numpy().repeat(class_dim, axis=0)
             one_hot = torch.cat([torch.from_numpy(oh), torch.from_numpy(one_hot)], dim=1).float()
         else:
-            one_hot = torch.from_numpy(one_hot)
+            one_hot = torch.from_numpy(one_hot).float()
         target_concat = torch.from_numpy(target.numpy().repeat(class_dim, axis=0)).float()
         
         data, target = data.to(device), target.to(device)
